requirements.txt
# Core Framework
Flask==3.0.0
Flask-Compress==1.14  # For gzip compression
Werkzeug==3.0.1
gunicorn==21.2.0

# Telegram Libraries
pyrogram==2.0.106
TgCrypto==1.2.5
aiogram==3.3.0

# Database
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
cryptography==41.0.7
requests==2.31.0

# Async Support
asyncio-mqtt==0.16.1
